feat: Add solution to delete middle node from a linked list
This commit adds a new solution to delete the middle node from a singly-linked list. The solution uses the two-pointer technique to find the middle node and removes it from the list. This is a useful operation when we want to remove a node from a linked list without having access to the head node.